Add Daily Regboot Cronjob
Requires the k8s deployment to be in the palworld namespace.

This launches a cronjob to run a shell script (defined in cm.yaml), which sends a broadcast notifying players that the server is saving. After 30 seconds, it does a formal backup using the established 'exec pod/podnamehere backup' command in the repo. After 30 seconds later, it sends a broadcast notifying players that the server will reboot in 60 seconds; 30 seconds; then each second for the last 10 seconds. It the runs the "Shutdown 1" command via rcon-cli, waits 30 additional seconds, then redeploys the application.

Currently, it is set to run at 9:30am UTC daily.

I'm newer to configuring such things, and will work to improve/cleanup a little bit. :)