chore(ci): Ignore hadolint false-positives
hadolint DL3029
> Do not use --platform flag with FROM

While this would be a good idea for single-arch images, on amd64 this image must still resolve to avoid build errors but is not used in the final image

---

hadolint DL3006
> Always tag the version of an image explicitly

Hadolint is unaware that the final FROM image is based on one of the two FROM's above, both of which are tagged. Therefore this is a false positive.